requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
SQLAlchemy==2.0.28
pytest==8.1.1
pytest-cov==4.1.0
httpx==0.27.0
tox==4.11.4
python-dotenv==1.0.0
